cleanup replay attack tests and refactor
